Fixes gpu -> cpu synchronization issues!!!
This has to be the most important commit of them all!!

The issue started when I was debugging mlpnet with mnist. I was getting
inconsistent results, and lately the mask was all zeros even when I was only
asking for a 90% pruning. I dug deeper into the code -- replicating it in the
notebook. It worked flawlessly there! But not using main.py.

Eventually I traced the issue to saving CUDA tensors to disk as well as
transfering weights and masks from GPU to CPU with `non_blocking=True`. It seems
that under the hood, the next operation progressed i.e. applying masks and
writing the weights to disk before the tensor could be fully copied over. It's a
classic example of over optimization using something which behaves differently
than assumed. Converting the non blocking transfer to a blocking one seems to
have fixed the issue.

The current version gives good results on the three architectures reported in
the CHITA paper.